requirements.txt
slack_bolt==1.18.0
revChatGPT==5.0.0
flask==2.3.2
python-dotenv==1.0.0
firebase-admin==6.1.0
firebase==4.0.1
jira==3.5.1